# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Puesto de Salud Bohórquez",
    "description": "API para gestionar pacientes, doctores, servicios, consultorios y citas.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:5000" }
  ],
...
  "paths": {
...
    "/servicios": {
      "get": {
        "tags": ["Servicios"],
        "summary": "Listar servicios",
        "responses": {
          "200": {
            "description": "Servicios listados",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Servicio" } } } }
          }
        }
      },

      "post": {
        "tags": ["Servicios"],
        "summary": "Crear un servicio",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ServicioCreate" } } }
        },
        "responses": {
          "201": { "description": "Servicio creado" }
        }
      }
    },
    "/servicios/{id}": {
      "put": {
        "tags": ["Servicios"],
        "summary": "Actualizar un servicio",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ServicioCreate" } } }
        },
        "responses": {
          "200": { "description": "Servicio actualizado" },
          "404": { "description": "Servicio no encontrado" }
        }
      },

      "delete": {
        "tags": ["Servicios"],
        "summary": "Eliminar un servicio",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Servicio eliminado" },
          "404": { "description": "Servicio no encontrado" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Servicio": {
        "type": "object",
        "properties": {
          "id_servicio": { "type": "integer" },
          "nombre": { "type": "string" },
          "duracion_minutos": { "type": "integer" },
          "tipo": { "type": "string" }
        }
      },
      "ServicioCreate": {
        "type": "object",
        "required": ["nombre", "duracion_minutos", "tipo"],
        "properties": {
          "nombre": { "type": "string" },
          "duracion_minutos": { "type": "integer" },
          "tipo": { "type": "string" }
        }
      },
...
    }
  }
}